fix: move RefreshCookiesResult to types directory
